Fetch erc20 balance
