+++ packages/controller/src/iframe/profile.ts
@@ -7,6 +7,7 @@ export type ProfileIFrameOptions = IFrameOptions<Profile> &
     address: string;
     username: string;
     indexerUrl: string;
+    rpcUrl: string;
   };
 
 export class ProfileIFrame extends IFrame<Profile> {
@@ -15,12 +16,14 @@ export class ProfileIFrame extends IFrame<Profile> {
     address,
     username,
     indexerUrl,
+    rpcUrl,
     ...iframeOptions
   }: ProfileIFrameOptions) {
     const _url = new URL(profileUrl ?? PROFILE_URL);
     _url.searchParams.set("address", encodeURIComponent(address));
     _url.searchParams.set("username", encodeURIComponent(username));
     _url.searchParams.set("indexerUrl", encodeURIComponent(indexerUrl));
+    _url.searchParams.set("rpcUrl", encodeURIComponent(rpcUrl));
 
     super({
       ...iframeOptions,
