+++ packages/profile/src/components/inventory.tsx
@@ -4,7 +4,9 @@ import {
   CardHeader,
   CardTitle,
   CopyAddress,
+  SpinnerIcon,
 } from "@cartridge/ui-next";
+import { erc20 } from "@cartridge/utils";
 import {
   LayoutContainer,
   LayoutContent,
@@ -12,9 +14,44 @@ import {
 } from "@/components/layout";
 import { useConnection } from "./provider/hooks";
 import { Navigation } from "./navigation";
+import { useEffect, useState } from "react";
 
 export function Inventory() {
-  const { username, address } = useConnection();
+  const { username, address, provider } = useConnection();
+  const [erc20s, setErc20s] = useState(ERC_20_TOKENS);
+  const [isFetching, setIsFetching] = useState(true);
+
+  useEffect(() => {
+    const id = setInterval(async () => {
+      setIsFetching(true);
+
+      const tokens = await Promise.all(
+        erc20s.map(async (t) => {
+          try {
+            const balance = await erc20.balanceOf({
+              accountAddress: address,
+              tokenAddress: t.address,
+              provider,
+            });
+
+            return {
+              ...t,
+              balance,
+            };
+          } catch (error) {
+            return { ...t, error };
+          }
+        }),
+      );
+
+      setErc20s(tokens);
+      setIsFetching(false);
+    }, 3000);
+
+    return () => {
+      clearInterval(id);
+    };
+  }, [address, erc20s, provider]);
 
   return (
     <LayoutContainer>
@@ -26,10 +63,11 @@ export function Inventory() {
 
       <LayoutContent className="pb-4">
         <Card>
-          <CardHeader>
+          <CardHeader className="h-10 flex flex-row items-center justify-between">
             <CardTitle>Token</CardTitle>
+            {isFetching && <SpinnerIcon className="animate-spin" />}
           </CardHeader>
-          {ERC_20_TOKENS.map((t, i) => (
+          {erc20s.map((t, i) => (
             <CardContent
               key={t.address + i}
               className="flex gap-x-1.5 items-center"
@@ -73,6 +111,6 @@ const ERC_20_TOKENS = [
       "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7",
     logoUrl:
       "https://imagedelivery.net/0xPAQaDtnQhBs8IzYRIlNg/e07829b7-0382-4e03-7ecd-a478c5aa9f00/logo",
-    balance: "0.01",
+    balance: "0",
   },
 ].flatMap((t) => [t, t, t]);
